% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{County irrigated acreage by category of irrigation}
\format{
A data frame with 4,232 rows and 6 variables:
\describe{
\item{County}{The name of the County, in title case ("McCormick")}
\item{Category}{Sprinkler, Low Flow / Drip, Gravity/Flood, or Total Irrigated Acreage}
\item{Subcategory}{Sprinkler, Drip, and Flood irrigation are subdivided into more specific categories.}
\item{Year}{1997-2000}
\item{Acres}{Number of acres irrigated}
\item{Source}{name of the extension agent who collected the data}
}
}
\source{
W Bryan Smith, Area Extension Agent, Clemson University
}
\usage{
categories
}
\description{
A dataset with the amount of acreage irrigated each year in each county
by each category of irrigation distribution system.
}
\details{
Data collected by county extension agents in South Carolina.
}
\references{
Irrigation Journal (2001)
}
\seealso{
\link{crops}, \link{irr_attributes}
}
\keyword{datasets}
