% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crops}
\alias{crops}
\title{County irrigated acreage by crop}
\format{
A data frame with 6,348 rows and 5 variables:
\describe{
\item{County}{The name of the County, in title case ("McCormick")}
\item{Crop}{Name of the irrigated crop, one of: Alfalfa, Apples, Avocado, Barley, Beans, Beans & Peas, Berries, Citrus, Corn, Cotton, Cranberries, Flower Seeds, Grains, Grapes, Grass Seeds, Hay Crops, Hops, Lettuce, Melons, Mint, Nursery, Nuts, Oats, Olives, Other, Pasture, Pasture/Hay Crop, Peanuts, Pecans, Pineapples, Potatoes, Rice, Small Fruits, Sorghum, Soybeans, Sugar Beets, Sugar Cane, Sunflowers, Sweet Potatoes, Tobacco, Tree Fruits, Turfgrass Sod, Vegetables, Wheat}
\item{Year}{1997-2000}
\item{Acres}{Number of acres irrigated}
\item{Source}{name of the extension agent who collected the data}
}
}
\source{
W Bryan Smith, Area Extension Agent, Clemson University
}
\usage{
crops
}
\description{
A dataset with the amount of acreage for each irrigated crop each year in each county.
}
\details{
Data collected by county extension agents in South Carolina.
}
\references{
Irrigation Journal (2001)
}
\seealso{
\link{categories}, \link{irr_attributes}
}
\keyword{datasets}
