% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuscirrigationsurvey-package.R
\docType{package}
\name{cuscirrigationsurvey-package}
\alias{cuscirrigationsurvey}
\alias{cuscirrigationsurvey-package}
\title{cuscirrigationsurvey: Irrigation Survey Data from Clemson University Extension, 1997-2000}
\description{
A dataset compiled from surveys conducted by Clemson Extension Field Agents from 1997-2000.
}
\references{
Irrigation Journal, 2001.

Charles Privette

W Bryan Smith
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/capellett/cuscirrigationsurvey}
  \item \url{https://capellett.github.io/cuscirrigationsurvey/}
  \item Report bugs at \url{https://github.com/capellett/cuscirrigationsurvey/issues}
}

}
\author{
\strong{Maintainer}: C A Pellett \email{c.alex.pellett@gmail.com}

Authors:
\itemize{
  \item W Bryan Smith \email{wsmth@clemson.edu}
}

Other contributors:
\itemize{
  \item State of South Carolina Department of Natural Resources [funder]
}

}
\keyword{internal}
