% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irr_attributes}
\alias{irr_attributes}
\title{County irrigation attributes}
\format{
A data frame with 2,392 rows and 6 variables:
\describe{
\item{County}{The name of the County, in title case ("McCormick")}
\item{Category}{Irrigation Power Unit Types, Irrigation Wells, or Pumps}
\item{Subcategory}{Different responses for each attribute}
\item{Year}{1997-2000}
\item{Value}{The total or percent response from each county}
\item{Source}{name of the extension agent who collected the data}
}
}
\source{
W Bryan Smith, Area Extension Agent, Clemson University
}
\usage{
irr_attributes
}
\description{
A dataset with the amount of acreage irrigated each year in each county
by each category of irrigation distribution system.
}
\details{
Data collected by county extension agents in South Carolina.
}
\references{
Irrigation Journal (2001)
}
\seealso{
\link{categories}, \link{crops}
}
\keyword{datasets}
